Chore/dependencies bump dec 2025 (#26)
For quite a long time we were not able to bump the dependencies, because
tsed had a breaking change which stopped support for CommonJS bundling.
In order to support ESM, we no longer could use `tsc` in the build
process, so I had to change the setup to use any serious bundler, like
tsup, and fix some Typescript issues to make it work.